Add files via upload
In this method , consider two arrays and two for loops. The first for loop will hold an element in the array and the second for loop will look for duplicate elements. If duplicate elements are found then , it will increment their frequencies and store them in another array. Also, a count variable is used to check if the element is previously counted or not .